third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/window-reuse-in-nested-browsing-contexts.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] synchronously navigate iframe with no initial src.
  assert_true: Not expecting event, but got load event expected true got false
[FAIL] after the first iframe load event, navigate iframe with no initial src.
  assert_true: Not expecting event, but got load event expected true got false
[FAIL] synchronously navigate iframe with initial src == "".
  assert_true: Not expecting event, but got load event expected true got false
[FAIL] after the first iframe load event, navigate iframe with initial src == "".
  assert_true: Not expecting event, but got load event expected true got false
[FAIL] synchronously navigate iframe with initial src == "about:blank".
  assert_true: Not expecting event, but got load event expected true got false
[FAIL] after the first iframe load event, navigate iframe with initial src == "about:blank".
  assert_true: Not expecting event, but got load event expected true got false
Harness: the test ran to completion.
